Migrate IDs from Long to BigInteger

+++ gcloud-java-compute/src/test/java/com/google/gcloud/compute/DiskTypeTest.java
@@ -20,9 +20,11 @@ import static org.junit.Assert.assertEquals;
 
 import org.junit.Test;
 
+import java.math.BigInteger;
+
 public class DiskTypeTest {
 
-  private static final Long ID = 42L;
+  private static final BigInteger ID = BigInteger.valueOf(42L);
   private static final String CREATION_TIMESTAMP = "2016-01-20T04:39:00.210-08:00";
   private static final String DESCRIPTION = "description";
   private static final String VALID_DISK_SIZE = "10GB-10TB";
