+++ gcloud-java-compute/src/test/java/com/google/gcloud/compute/MachineTypeTest.java
@@ -22,11 +22,12 @@ import com.google.common.collect.ImmutableList;
 
 import org.junit.Test;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public class MachineTypeTest {
 
-  private static final Long ID = 42L;
+  private static final BigInteger ID = BigInteger.valueOf(42L);
   private static final String CREATION_TIMESTAMP = "2016-01-20T04:39:00.210-08:00";
   private static final String DESCRIPTION = "description";
   private static final MachineTypeId MACHINE_TYPE_ID = MachineTypeId.of("project", "zone", "type");
